move slice and reduce
